+++ torch/csrc/jit/pickle.h
@@ -3,7 +3,7 @@
 #include <torch/csrc/WindowsTorchApiMacro.h>
 #include <ATen/core/ivalue.h>
 #include <torch/csrc/jit/pickler.h>
-
+#include <torch/csrc/jit/unpickler.h>
 
 namespace torch {
 namespace jit {
